Rest end point issue resolved, but need to check with actual data

+++ src/main/java/com/advt/service/AdvertisementUserService.java
@@ -51,8 +51,7 @@ public class AdvertisementUserService {
         return targetedAdvertisementPlanRepository.findAvailablePlansByLocationId(locationId);
     }
 
-	public Long getTotalSubscribersByLocationId(Integer locationId) {
-        return userSubscriptionRepository.countSubscribersByLocationId(locationId);
+	public Long getTotalSubscribersByLocationId(int locationId) {
+        return userSubscriptionRepository.countSubscribersByLocationId((long) locationId);
     }
-
 }
